fix: delay in useHover is not working as expected
In practive when we say "delay the mouse enter" we are expecting callbacks to trigger after a certain time **and it should not trigger if we leave the element within that delay**, however current implement will trigger callback after a certain delay even we have already left that element.

To fix this, we have to manage a timer in hook and cancel all timeouts when hover state changes.